move stuff to own file
